requirements.txt
Django==1.9.5
boto3==1.3.0
MySQL-python==1.2.5
uwsgi
djangorestframework
markdown
django-filter
django-crispy-forms
requests
